pyproject.toml
[project]
name = "pay"
version = "0.1.0"
description = "Payment system with FastAPI"
authors = [
    { name = "Alan" },
]
requires-python = ">=3.13"
dependencies = [
    "asyncer>=0.0.8",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.2.2",
    "loguru>=0.7.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "qrcode>=8.2",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
